+++ bindings/python/src/lib.rs
@@ -42,6 +42,9 @@ impl From<InlineErrorWrapper> for PyErr {
             rust_inline::InlineError::ParseError(message) => {
                 InlineError::new_err(message.to_string())
             }
+            rust_inline::InlineError::MissingStyleSheet { .. } => {
+                InlineError::new_err(error.0.to_string())
+            }
         }
     }
 }